requirements-api.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
Pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
omegaconf==2.3.0
PyYAML==6.0.2
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
diffusers==0.29.2
transformers==4.44.2
accelerate==0.34.2

